# api-docs-cuentas-movimientos.json
{"openapi":"3.0.1","info":{"title":"API de Cuentas y Movimientos","description":"Documentación de la API para la gestión de cuentas y movimientos","version":"V1"},"servers":[{"url":"http://localhost:8081","description":"Generated server url"}],"tags":[{"name":"Cuentas","description":"Operaciones relacionadas con las cuentas"},{"name":"Movimientos","description":"Operaciones relacionadas con los movimientos"},{"name":"Reportes","description":"Operaciones relacionadas con los reportes de movimientos"}],"paths":{"/api/v1/movimientos/{id}":{"get":{"tags":["Movimientos"],"summary":"Obtener un movimiento por ID","operationId":"obtenerMovimientoPorId","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"404":{"description":"Movimiento no encontrado","content":{"*/*":{"schema":{"$ref":"#/components/schemas/MovimientoDTO"}}}},"200":{"description":"Movimiento encontrado","content":{"*/*":{"schema":{"$ref":"#/components/schemas/MovimientoDTO"}}}}}},"put":{"tags":["Movimientos"],"summary":"Actualizar un movimiento","operationId":"actualizarMovimiento","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/MovimientoDTO"}}},"required":true},"responses":{"404":{"description":"Movimiento no encontrado","content":{"*/*":{"schema":{"$ref":"#/components/schemas/MovimientoDTO"}}}},"400":{"description":"Error en la validación de los datos del movimiento","content":{"*/*":{"schema":{"$ref":"#/components/schemas/MovimientoDTO"}}}},"200":{"description":"Movimiento actualizado exitosamente","content":{"*/*":{"schema":{"$ref":"#/components/schemas/MovimientoDTO"}}}}}},"delete":{"tags":["Movimientos"],"summary":"Eliminar un movimiento","operationId":"eliminarMovimiento","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"404":{"description":"Movimiento no encontrado"},"204":{"description":"Movimiento eliminado exitosamente"}}}},"/api/v1/cuentas/{id}":{"get":{"tags":["Cuentas"],"summary":"Obtener una cuenta por ID","operationId":"obtenerCuentaPorId","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"404":{"description":"Cuenta no encontrada","content":{"*/*":{"schema":{"$ref":"#/components/schemas/CuentaDTO"}}}},"200":{"description":"Cuenta encontrada","content":{"*/*":{"schema":{"$ref":"#/components/schemas/CuentaDTO"}}}}}},"put":{"tags":["Cuentas"],"summary":"Actualizar una cuenta","operationId":"actualizarCuenta","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CuentaDTO"}}},"required":true},"responses":{"404":{"description":"Cuenta no encontrada","content":{"*/*":{"schema":{"$ref":"#/components/schemas/CuentaDTO"}}}},"200":{"description":"Cuenta actualizada exitosamente","content":{"*/*":{"schema":{"$ref":"#/components/schemas/CuentaDTO"}}}},"400":{"description":"Error en la validación de los datos de la cuenta","content":{"*/*":{"schema":{"$ref":"#/components/schemas/CuentaDTO"}}}}}},"delete":{"tags":["Cuentas"],"summary":"Eliminar una cuenta","operationId":"eliminarCuenta","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"404":{"description":"Cuenta no encontrada"},"204":{"description":"Cuenta eliminada exitosamente"}}}},"/api/v1/movimientos":{"get":{"tags":["Movimientos"],"summary":"Obtener todos los movimientos","operationId":"obtenerMovimientos","responses":{"200":{"description":"Operación exitosa","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/MovimientoDTO"}}}}},"204":{"description":"No se encontraron movimientos","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/MovimientoDTO"}}}}}}},"post":{"tags":["Movimientos"],"summary":"Crear un nuevo movimiento","operationId":"crearMovimiento","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/MovimientoDTO"}}},"required":true},"responses":{"400":{"description":"Error en la validación de los datos del movimiento","content":{"*/*":{"schema":{"$ref":"#/components/schemas/MovimientoDTO"}}}},"201":{"description":"Movimiento creado exitosamente","content":{"*/*":{"schema":{"$ref":"#/components/schemas/MovimientoDTO"}}}}}}},"/api/v1/cuentas":{"get":{"tags":["Cuentas"],"summary":"Obtener todas las cuentas","operationId":"obtenerCuentas","responses":{"204":{"description":"No se encontraron cuentas","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/CuentaDTO"}}}}},"200":{"description":"Operación exitosa","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/CuentaDTO"}}}}}}},"post":{"tags":["Cuentas"],"summary":"Crear una nueva cuenta","operationId":"crearCuenta","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CuentaDTO"}}},"required":true},"responses":{"201":{"description":"Cuenta creada exitosamente","content":{"*/*":{"schema":{"$ref":"#/components/schemas/CuentaDTO"}}}},"400":{"description":"Error en la validación de los datos de la cuenta","content":{"*/*":{"schema":{"$ref":"#/components/schemas/CuentaDTO"}}}}}}},"/api/v1/reportes":{"get":{"tags":["Reportes"],"summary":"Generar reporte de movimientos","operationId":"obtenerReporteMovimientos","parameters":[{"name":"clienteId","in":"query","required":true,"schema":{"type":"integer","format":"int64"}},{"name":"fechaInicio","in":"query","required":true,"schema":{"type":"string","format":"date"}},{"name":"fechaFin","in":"query","required":true,"schema":{"type":"string","format":"date"}}],"responses":{"404":{"description":"No se encontraron movimientos para el rango de fechas","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/ReporteMovimientoDTO"}}}}},"200":{"description":"Reporte generado exitosamente","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/ReporteMovimientoDTO"}}}}}}}}},"components":{"schemas":{"MovimientoDTO":{"required":["numeroCuenta","tipo","valor"],"type":"object","properties":{"id":{"type":"integer","format":"int64"},"uniqueId":{"type":"string"},"fecha":{"type":"string","format":"date-time"},"tipo":{"maxLength":3,"minLength":0,"type":"string"},"valor":{"minimum":0.0,"exclusiveMinimum":true,"type":"number"},"numeroCuenta":{"type":"string"},"saldoInicial":{"type":"number"},"saldoDisponible":{"type":"number"}}},"CuentaDTO":{"required":["clienteId","estado","numeroCuenta","saldoInicial","tipo"],"type":"object","properties":{"id":{"type":"integer","format":"int64"},"uniqueId":{"type":"string"},"numeroCuenta":{"maxLength":20,"minLength":0,"type":"string"},"tipo":{"maxLength":3,"minLength":0,"type":"string"},"saldoInicial":{"type":"number"},"estado":{"type":"boolean"},"clienteId":{"type":"integer","format":"int64"}}},"ReporteMovimientoDTO":{"type":"object","properties":{"fecha":{"type":"string","format":"date-time"},"cliente":{"type":"string"},"numeroCuenta":{"type":"string"},"tipoCuenta":{"type":"string"},"tipoMovimiento":{"type":"string"},"saldoInicial":{"type":"number"},"valorMovimiento":{"type":"number"},"saldoDisponible":{"type":"number"}}}}}}